agents/manual_assistants/requirements.txt
docutils>=0.20.1
Jinja2
openai
PyYAML
requests
